+++ src/components/scenes/Timeline.vue
@@ -3,9 +3,10 @@ import {onMounted, ref} from 'vue';
 
 import {useScenesStore} from '../../stores/scenesStore.ts';
 
-import {Action} from '../../types/scene.ts';
+import {Action, JumpAction, MessageAction, RawAction} from '../../types/scene.ts';
 
-import ActionRaw from './actions/ActionRaw.vue';
+import TimelineAction from './actions/Action.vue';
+import DropZone from './DropZone.vue';
 import NewActions from './NewActions.vue';
 
 const props = defineProps<{
@@ -14,8 +15,7 @@ const props = defineProps<{
 
 const actions = ref<Action[]>([]);
 const selectedAction = ref<Action['type'] | null>(null);
-const dropZone = ref<HTMLDivElement | null>(null);
-const dropZoneSelected = ref<boolean>(false);
+const orderSelected = ref<number | null>(null);
 
 function init(): void {
     actions.value = [];
@@ -25,50 +25,65 @@ function init(): void {
 }
 
 function addAction(): void {
-    if (!selectedAction.value) return;
+    console.log('addAction', selectedAction.value, orderSelected.value);
+    if (selectedAction.value === null || orderSelected.value === null) return;
+
     const actionName = selectedAction.value;
-    const newAction: Partial<Action> = {};
+    const newAction = getAction(actionName);
+    if (!newAction) return;
+
+    useScenesStore().createAction(actionName, newAction, props.sceneId, orderSelected.value);
+    removeOrderSelected(orderSelected.value);
+    init();
+}
+
+function getAction(actionName: Action['type']): Action | null {
+    let newAction: Partial<Action> = {};
 
     switch (actionName) {
+        case 'message':
+            const messageAction = newAction as MessageAction;
+            messageAction['message'] = '';
+            messageAction['character'] = null;
+            newAction = messageAction;
+            break;
+        case 'jump':
+            const jumpAction = newAction as JumpAction;
+            jumpAction['sceneId'] = null;
+            newAction = jumpAction;
+            break;
         case 'raw':
-            newAction['code'] = '';
+            const rawAction = newAction as RawAction;
+            rawAction['code'] = '';
+            newAction = rawAction;
             break;
         default:
             console.error('Unknown action type: ', actionName);
-            return;
+            return null;
     }
 
-    useScenesStore().createAction(actionName, newAction, props.sceneId);
-    init();
+    return newAction as Action;
 }
 
 function startDrag(name: Action['type']): void {
     selectedAction.value = name;
-    dropZone.value?.classList.add('drop-zone--visible');
 }
 
 function endDrag(): void {
-    dropZone.value?.classList.remove('drop-zone--visible');
-    dropZone.value?.classList.remove('drop-zone--drag-hover');
-
-    if (dropZoneSelected.value) {
+    if (selectedAction.value !== null || orderSelected.value !== null) {
         addAction();
     }
-
-    dropZoneSelected.value = false;
+    selectedAction.value = null;
 }
 
-function dragEnter(): void {
-    if (dropZone.value?.classList.contains('drop-zone--drag-hover')) return;
-    dropZone.value?.classList.add('drop-zone--drag-hover');
-    dropZoneSelected.value = true;
+function setOrderSelected(order: number): void {
+    orderSelected.value = order;
 }
 
-function dragLeave(): void {
-    setTimeout(() => {
-        dropZone.value?.classList.remove('drop-zone--drag-hover');
-        dropZoneSelected.value = false;
-    }, 100);
+function removeOrderSelected(order: number): void {
+    if (orderSelected.value === order) {
+        orderSelected.value = null;
+    }
 }
 
 onMounted(init);
@@ -77,21 +92,35 @@ onMounted(init);
 <template>
   <div class="timeline">
     <div class="timeline__content">
-      <div
+      <TransitionGroup
         v-for="action in actions"
         :key="action._id"
         class="action"
+        tag="div"
       >
-        <ActionRaw
-          v-if="action.type === 'raw'"
-          :action="action"
+        <DropZone
+          :key="action._id + '-drop'"
+          :actions="actions"
+          :is-visible="selectedAction !== null || orderSelected === action._order - 1"
+          :order="action._order - .5"
+          :selected-action="selectedAction"
+          class="drop-zone--not-last"
+          @set-drag-order="setOrderSelected"
+          @remove-drag-order="removeOrderSelected"
         />
-      </div>
-      <div
-        ref="dropZone"
-        class="drop-zone"
-        @dragenter="dragEnter"
-        @dragleave="dragLeave"
+        <TimelineAction
+          :key="action._id"
+          :action="action"
+          @updated="init"
+        />
+      </TransitionGroup>
+      <DropZone
+        :actions="actions"
+        :is-visible="selectedAction !== null || actions.length === 0"
+        :order="actions.length"
+        :selected-action="selectedAction"
+        @set-drag-order="setOrderSelected"
+        @remove-drag-order="removeOrderSelected"
       />
     </div>
     <NewActions
@@ -106,10 +135,15 @@ onMounted(init);
     display: grid;
     gap: .5rem;
     grid-template-columns: 1fr 40px;
+    overflow-y: hidden;
+    height: calc(100vh - 8.3rem);
 
     &__content {
         display: grid;
-        gap: .25rem;
+        align-content: start;
+        gap: .5rem;
+        overflow-y: auto;
+        padding: .5rem .5rem 50px;
     }
 }
 
@@ -120,6 +154,8 @@ onMounted(init);
     border-radius: .5rem;
     position: relative;
     height: 10px;
+    opacity: .5;
+    width: calc(100% - calc(60px + 1rem));
 
     &--visible {
         border: 2px dashed var(--color-accent);
@@ -127,6 +163,11 @@ onMounted(init);
 
     &--drag-hover {
         height: 35px !important;
+        opacity: 1;
+    }
+
+    &--not-last {
+        margin-bottom: .5rem;
     }
 }
 </style>