+++ src/utils/pamyu.ts
@@ -4,7 +4,7 @@ import {useCharacterStore} from '../stores/characterStore.ts';
 import {useScenesStore} from '../stores/scenesStore.ts';
 
 import {Character} from '../types/character.ts';
-import {Action, RawAction} from '../types/scene.ts';
+import {Action, JumpAction, MessageAction, RawAction} from '../types/scene.ts';
 
 import {path} from '../main';
 
@@ -37,9 +37,10 @@ export function getCharactersScript(): string {
 export function updateScenesScipts(): void {
     console.log('📂 Updating scenes scripts');
 
+    // TODO: Save scenes and check if they have changed before last update
     useScenesStore().getScenes.forEach((scene) => {
         let sceneString = `${warnMessage}label ${scene._id}:\n`;
-        const indent = getIndent(1);
+        const indent = getIndent();
         const actionsLines = getActionsLines(useScenesStore().getAllActionsOfScene(scene._id));
         sceneString += indent + actionsLines.join(`\n${indent}`);
 
@@ -64,6 +65,12 @@ function getActionLines(action: Action): string[] {
     const lines: string[] = [];
 
     switch (action.type) {
+        case 'message':
+            lines.push(...getMessageActionLines(action));
+            break;
+        case 'jump':
+            lines.push(...getJumpActionLines(action));
+            break;
         case 'raw':
             lines.push(...getRawActionLines(action));
             break;
@@ -73,9 +80,50 @@ function getActionLines(action: Action): string[] {
 }
 
 function getRawActionLines(action: RawAction): string[] {
+    if (action.code.trim().length === 0) return [];
     const lines: string[] = [];
 
     lines.push(...action.code.split('\n'));
 
     return lines;
+}
+
+function getJumpActionLines(action: JumpAction): string[] {
+    const lines: string[] = [];
+
+    if (action.sceneId) {
+        lines.push(`jump ${action.sceneId}`);
+    }
+
+    return lines;
+}
+
+function getMessageActionLines(action: MessageAction): string[] {
+    if (!action.character && action.message.length === 0) return [];
+
+    const lines: string[] = [];
+    const characterString = getCharacterString(action);
+
+    if (action.message.includes('\n')) {
+        lines.push(`${characterString}"""`);
+        action.message.split('\n').forEach((line) => lines.push(getIndent() + line));
+        lines.push('"""');
+    } else {
+        lines.push(`${characterString}"${action.message}"`);
+    }
+
+    return lines;
+}
+
+function getCharacterString(action: MessageAction): string {
+    let characterString = '';
+
+    if (action.character) {
+        const character = useCharacterStore().getCharacterById(action.character);
+        if (character) {
+            characterString = character._id + ' ';
+        }
+    }
+
+    return characterString;
 }