+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## Alpha 10
+
+✨ Actions inprovements, can be deleted and ordered.
+
+🧶 Add "Message" action.
+
+🧶 Add "Jump" action.
+
+🎨 Add some quality of life improvements.
+
+🎨 Add some animations.
+
+🗯️ Added new translations:
+
+- 🐪 Arabic
+- 🍛 Bengali
+- 🌭 German
+- 💃 Spanish
+- 🕌 Hindi
+- 🌸 Japanese
+- 🍷 Portuguese
+- ❄️ Russian
+- 🏮 Chinese
+
+---
+
 ## Alpha 9
 
 ✨ Add the scene section.
@@ -14,6 +40,8 @@
 
 🚀 Now using better semantic versioning and release-it.
 
+---
+
 ## Alpha 8
 
 ✨ Can now launch the game and Ren'Py directly.
@@ -27,6 +55,8 @@
 🎨 Refinement of the overall design.
 
 🎨 New theme selector.
+
+---
 
 ## Alpha 7